fixed training hooks and cleaned
